requirements.txt
ebooklib
google-generativeai
python-dotenv
BeautifulSoup4